% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggeo-package.R
\docType{package}
\name{ggeo-package}
\alias{ggeo}
\alias{ggeo-package}
\title{ggeo: Provide Helpers and Themes for ggplot}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package provides helper functions for ggplot graphs and maps.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ggeo.nenuial.org}
}

}
\author{
\strong{Maintainer}: Pascal Burkhard \email{pascal.burkhard@gmail.com} (\href{https://orcid.org/0009-0001-5504-5084}{ORCID})

}
\keyword{internal}
