% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_utils.R
\name{ggeo_capitalize_title}
\alias{ggeo_capitalize_title}
\title{Copitalize plot title}
\usage{
ggeo_capitalize_title(plot)
}
\arguments{
\item{plot}{A ggplot2 object}
}
\value{
A ggplot2 object
}
\description{
Copitalize plot title
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
ggplot(aes(x = speed, y = dist), data = cars) +
  geom_point() +
  labs(title = "Fast cars !") +
  ggeo_remove_title()
\dontshow{\}) # examplesIf}
}
