% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{ggeo_coord}
\alias{ggeo_coord}
\title{Return sf coord for given crs code}
\usage{
ggeo_coord(proj, ...)
}
\arguments{
\item{proj}{A string for the projection code}

\item{...}{Arguments passed to ggplot2::coord_sf}
}
\value{
A ggplot2 coord object
}
\description{
Return sf coord for given crs code
}
