% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharter_themes.R
\name{ggeo_create_highchart_theme}
\alias{ggeo_create_highchart_theme}
\title{Function factory to create a highcharter theme}
\usage{
ggeo_create_highchart_theme(
  base_colors,
  palette,
  web_theme = FALSE,
  title_font_size = if (!web_theme) "52px" else NULL,
  title_font_weight = if (!web_theme) "400" else NULL,
  subtitle_font_size = if (!web_theme) "38px" else NULL,
  subtitle_font_weight = if (!web_theme) "300" else NULL
)
}
\arguments{
\item{base_colors}{A list with the main, axis and title colors}

\item{palette}{A color palette}

\item{web_theme}{A boolean to indicate if the theme is for the web}

\item{title_font_size}{The title font size}

\item{title_font_weight}{The title font weight}

\item{subtitle_font_size}{The subtitle font size}

\item{subtitle_font_weight}{The subtitle font weight}
}
\value{
A highcharter theme object
}
\description{
Function factory to create a highcharter theme
}
\keyword{internal}
