% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{ggeo_file}
\alias{ggeo_file}
\title{A shortcut function for the system.file to the ggeo package}
\usage{
ggeo_file(...)
}
\value{
A path
}
\description{
A shortcut function for the system.file to the ggeo package
}
