% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{ggeo_install_fonts_linux}
\alias{ggeo_install_fonts_linux}
\title{Install fonts on Linux}
\usage{
ggeo_install_fonts_linux()
}
\value{
Move fonts to Linux font folders
}
\description{
Install fonts on Linux
}
