% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{ggeo_label_abs_percent}
\alias{ggeo_label_abs_percent}
\title{Format labels for relative pyramids}
\usage{
ggeo_label_abs_percent(x)
}
\arguments{
\item{x}{Number to format}
}
\value{
A formatted expression
}
\description{
Format labels for relative pyramids
}
