% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{ggeo_label_pyramid}
\alias{ggeo_label_pyramid}
\title{Format labels for pyramids}
\usage{
ggeo_label_pyramid(x)
}
\arguments{
\item{x}{Number to format}
}
\value{
A formatted expression
}
\description{
Format labels for pyramids
}
