% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{ggeo_label_sci_10}
\alias{ggeo_label_sci_10}
\title{Format labels using scientific notation}
\usage{
ggeo_label_sci_10(x)
}
\arguments{
\item{x}{Number to format}
}
\value{
A formatted expression
}
\description{
Format labels using scientific notation
}
