% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{ggeo_remove_title}
\alias{ggeo_remove_title}
\title{Remove plot title}
\usage{
ggeo_remove_title(plot)
}
\arguments{
\item{plot}{A ggplot2 object}
}
\value{
A ggplot2 object
}
\description{
Remove plot title
}
