% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_utils.R
\name{ggeo_save}
\alias{ggeo_save}
\title{Save function}
\usage{
ggeo_save(
  plot,
  filename,
  width = geotools::gtl_options("plot_standard_width"),
  height = geotools::gtl_options("plot_standard_height"),
  dpi = 72,
  units = "cm",
  ...
)
}
\arguments{
\item{plot}{The ggplot2 object to save (should be piped in ;)}

\item{filename}{Path for filename (with extension!)}

\item{width}{The plot width.
Defaults to keynote width (\code{geotools::gtl_options("plot_standard_width")}).
For full slide width use \code{geotools::gtl_options("plot_full_width")}.}

\item{height}{The plot height.
Defaults to keynote height (\code{geotools::gtl_options("plot_standard_height")}).
For full slide height use \code{geotools::gtl_options("plot_full_height")}.}

\item{dpi}{The DPI. Default is 72.}

\item{units}{Units. Default is cm.}

\item{...}{
  Arguments passed on to \code{\link[ggplot2:ggsave]{ggplot2::ggsave}}
  \describe{
    \item{\code{device}}{Device to use. Can either be a device function
(e.g. \link{png}), or one of "eps", "ps", "tex" (pictex),
"pdf", "jpeg", "tiff", "png", "bmp", "svg" or "wmf" (windows only). If
\code{NULL} (default), the device is guessed based on the \code{filename} extension.}
    \item{\code{scale}}{Multiplicative scaling factor.}
    \item{\code{limitsize}}{When \code{TRUE} (the default), \code{ggsave()} will not
save images larger than 50x50 inches, to prevent the common error of
specifying dimensions in pixels.}
    \item{\code{bg}}{Background colour. If \code{NULL}, uses the \code{plot.background} fill value
from the plot theme.}
    \item{\code{create.dir}}{Whether to create new directories if a non-existing
directory is specified in the \code{filename} or \code{path} (\code{TRUE}) or return an
error (\code{FALSE}, default). If \code{FALSE} and run in an interactive session,
a prompt will appear asking to create a new directory when necessary.}
  }}
}
\description{
Save function
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
cars |>
  ggplot2::ggplot(ggplot2::aes(x = speed, y = dist)) +
  ggplot2::geom_point() -> simple_plot

ggeo_save(simple_plot, "simple_plot.png")
\dontshow{\}) # examplesIf}
}
