% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{ggeoformat_pyramid_pop}
\alias{ggeoformat_pyramid_pop}
\title{Format population labels on pyramids}
\usage{
ggeoformat_pyramid_pop(x, ...)
}
\arguments{
\item{x}{The label}

\item{...}{For compatibility}
}
\value{
Absolute
}
\description{
DEPRECATED
}
