% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{ggeopal_center}
\alias{ggeopal_center}
\title{Return a color palette with option to center diverging palettes}
\usage{
ggeopal_center(n, center, params)
}
\arguments{
\item{n}{Number of colors}

\item{center}{Where is the center (counted from the left).
Use -1 for non diverging palettes.}

\item{params}{Parameters to pass on to paletteer}
}
\value{
A vector of colors
}
\description{
Return a color palette with option to center diverging palettes
}
