% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{ggeopal_chooser}
\alias{ggeopal_chooser}
\title{Return a paletteer palette depending on}
\usage{
ggeopal_chooser(n, params)
}
\arguments{
\item{n}{Number of colors}

\item{params}{Parameters to pass on to paletteer
Must be a list with:
\itemize{
\item type: the type of palette, on of \code{cont}, \code{dis} or \code{dyn}
\item palette: the {paletteer} palette
\item dir: the palette direction
}}
}
\value{
A vector of colors
}
\description{
Return a paletteer palette depending on
}
