% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{ggeopal_hex_to_hcl}
\alias{ggeopal_hex_to_hcl}
\title{Generate colors from a HEX color using HCL palette}
\usage{
ggeopal_hex_to_hcl(hex, n = 4)
}
\arguments{
\item{hex}{A string HEX color}

\item{n}{Number of colors wanted}
}
\value{
A string vector of HEX colors
}
\description{
Generate colors from a HEX color using HCL palette
}
\examples{

ggeopal_hex_to_hcl("#222b4c") |>
  prismatic::color()

ggeopal_hex_to_hcl("#222b4c", 6) |>
  prismatic::color()

}
