% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_color.R
\name{ggeopal_pal_to_gradient}
\alias{ggeopal_pal_to_gradient}
\title{Create gradient for palette colors}
\usage{
ggeopal_pal_to_gradient(pal)
}
\arguments{
\item{pal}{A vector of colors}
}
\value{
Print color vectors
}
\description{
Create gradient for palette colors
}
