% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{ggeopal_to_gradient}
\alias{ggeopal_to_gradient}
\title{Create gradient for palette colors}
\usage{
ggeopal_to_gradient(pal)
}
\arguments{
\item{pal}{A vector of colors}
}
\value{
Print color vectors
}
\description{
Create gradient for palette colors
}
\examples{
ggeopal_to_gradient("#dd4123")
ggeopal_to_gradient("#ee950b")
ggeopal_to_gradient("#0b6b8b")

}
