% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_utils.R
\name{ggeosave}
\alias{ggeosave}
\title{Save plots to out folder}
\usage{
ggeosave(
  filename,
  ...,
  format = c(NA, "keynote"),
  width = 63.5,
  height = 28.57,
  dpi = 72,
  units = "cm",
  extension = "pdf",
  device = grDevices::cairo_pdf
)
}
\arguments{
\item{filename}{Name of the output file}

\item{...}{Other arguments passed on to the graphics device function,
as specified by \code{device}.}

\item{format}{Optional: one of "keynote"}

\item{width}{Optional: width of output plot}

\item{height}{Optional: height of output plot}

\item{dpi}{Plot resolution. Also accepts a string input: "retina" (320),
"print" (300), or "screen" (72). Applies only to raster output types.}

\item{units}{Optional: units of output plot dimensions}

\item{extension}{Optional: output file extension}

\item{device}{Optional: output device}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\keyword{internal}
