% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{ggeotheme}
\alias{ggeotheme}
\title{ggeo theme}
\usage{
ggeotheme(
  theme = c("ghibli_mononoke", "islamic_samarquand", "pomological_green",
    "pomological_red", "nord_blue"),
  ...,
  mode = c("light", "dark"),
  base = ggplot2::theme_bw()
)
}
\arguments{
\item{theme}{Name of the theme to use}

\item{mode}{One of light or dark}

\item{base}{Base ggplot2 theme to use}

\item{horizontal}{\code{logical}. Horizontal axis lines?}
}
\value{
An object of class \code{\link[ggplot2]{theme}()}.
}
\description{
ggeo theme
}
\concept{themes nord_dark}
