% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharter_themes.R
\name{hc_purple_theme}
\alias{hc_purple_theme}
\title{Highchart Theme: Purple
Add a purple theme to a highchart}
\usage{
hc_purple_theme(hc)
}
\arguments{
\item{hc}{A highchart object}
}
\value{
A highchart object
}
\description{
Highchart Theme: Purple
Add a purple theme to a highchart
}
