% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharter_themes.R
\name{hc_samarqand_theme}
\alias{hc_samarqand_theme}
\title{Highchart Theme: Samarqand
Add a samarqand theme to a highchart}
\usage{
hc_samarqand_theme(hc)
}
\arguments{
\item{hc}{A highchart object}
}
\value{
A highchart object
}
\description{
Highchart Theme: Samarqand
Add a samarqand theme to a highchart
}
