% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharter_themes.R
\name{hc_web_theme}
\alias{hc_web_theme}
\title{Highchart Theme: Web
Add a web theme to a highchart}
\usage{
hc_web_theme(hc)
}
\arguments{
\item{hc}{A highchart object}
}
\value{
A highchart object
}
\description{
Highchart Theme: Web
Add a web theme to a highchart
}
