% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_palettes.R
\name{palette_chooser}
\alias{palette_chooser}
\title{Return a paletteer palette depending on}
\usage{
palette_chooser(n, params)
}
\arguments{
\item{n}{Number of colors}

\item{params}{Parameters to pass on to paletteer}
}
\value{
A vector of colors
}
\description{
Return a paletteer palette depending on
}
