% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{return_palette}
\alias{return_palette}
\title{Return a color palette with option to center diverging palettes}
\usage{
return_palette(n, center, params)
}
\arguments{
\item{n}{Number of colors}

\item{center}{Where is the center (counted from the left). Use -1 for non diverging palettes.}

\item{params}{Parameters to pass on to paletteer}
}
\value{
A vector of colors
}
\description{
Return a color palette with option to center diverging palettes
}
