% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggeo-package.R
\docType{package}
\name{ggeo-package}
\alias{ggeo}
\alias{ggeo-package}
\title{ggeo: Provide Helpers and Themes for ggplot}
\description{
This package provides helper functions for ggplot graphs and maps.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ggeo.nenuial.org}
}

}
\author{
\strong{Maintainer}: Pascal Burkhard \email{pascal.burkhard@gmail.com}

}
\keyword{internal}
