% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{ggeo_capitalize_title}
\alias{ggeo_capitalize_title}
\title{Copitalize plot title}
\usage{
ggeo_capitalize_title(plot)
}
\arguments{
\item{plot}{A ggplot2 object}
}
\value{
A ggplot2 object
}
\description{
Copitalize plot title
}
