% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_scales.R
\name{ggeo_label_sci_10}
\alias{ggeo_label_sci_10}
\alias{ggeo_label_pyramid}
\alias{ggeo_label_abs_percent}
\title{Format labels}
\usage{
ggeo_label_sci_10(x)

ggeo_label_pyramid(x)

ggeo_label_abs_percent(x)
}
\arguments{
\item{x}{Number to format}
}
\value{
A formatted string for the scales
}
\description{
These functions extend the {\href{https://scales.r-lib.org}{scales}}
package and allow formatting labels.
}
\details{
\code{\link[=ggeo_label_sci_10]{ggeo_label_sci_10()}} is used to format numbers with a clean
scientific format using a multiplier and not the \emph{ugly} notation
using the letter \emph{e}.

\code{\link[=ggeo_label_pyramid]{ggeo_label_pyramid()}} is used for population pyramids with
absolute numbers. It formats the absolute number using
\code{\link[=ggeo_label_sci_10]{ggeo_label_sci_10()}}.

\code{\link[=ggeo_label_abs_percent]{ggeo_label_abs_percent()}} is also used for population pyramids
but with relative numbers (percents). It uses \code{\link[scales:percent_format]{scales::percent()}}
and absolute numbers.
}
