% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{ggeo_remove_breaks}
\alias{ggeo_remove_breaks}
\title{Remove specific breaks}
\usage{
ggeo_remove_breaks(original_func, remove_list = list())
}
\arguments{
\item{original_func}{The break function}

\item{remove_list}{The values to remove from the breaks}
}
\value{
A list
}
\description{
Remove specific breaks
}
