% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_save.R
\name{ggeo_save}
\alias{ggeo_save}
\title{Save function}
\usage{
ggeo_save(
  plot,
  filename,
  width = geotools::gtl_options("plot_standard_width"),
  height = geotools::gtl_options("plot_standard_height"),
  dpi = 72,
  units = "cm",
  ...
)
}
\arguments{
\item{plot}{The ggplot2 object to save (should by piped in ;)}

\item{filename}{Path for filename (with extension!)}

\item{width}{Width. Defaults to keynote width (full = geotools::gtl_options("plot_full_width"))}

\item{height}{Height. Defaults to keynote height (full = geotools::gtl_options("plot_full_width"))}

\item{dpi}{DPI. Defaults to 72.}

\item{units}{Units. Defaults to cm.}

\item{...}{Arguments to pass to ggplot2::ggsave}
}
\description{
Save function
}
