% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_scales.R
\name{ggeoformat_pyramid_pop}
\alias{ggeoformat_pyramid_pop}
\title{Format population labels on pyramids}
\usage{
ggeoformat_pyramid_pop(x, ...)
}
\arguments{
\item{x}{The label}

\item{...}{For compatibility}
}
\value{
Absolute
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\keyword{internal}
