% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_color.R
\name{ggeopal_hex_to_hcl}
\alias{ggeopal_hex_to_hcl}
\title{Generate four colors from a HEX color using HCL palette}
\usage{
ggeopal_hex_to_hcl(hex, n = 4)
}
\arguments{
\item{hex}{A string HEX color}
}
\value{
A string vector of HEX colors
}
\description{
Generate four colors from a HEX color using HCL palette
}
