% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{ggeotheme}
\alias{ggeotheme}
\title{ggeo theme}
\usage{
ggeotheme(
  theme = c("ghibli_mononoke", "islamic_samarquand", "pomological_green",
    "pomological_red", "nord_blue", "swiss_red", "doc", "oc_exams"),
  main = "main",
  plot = "plot",
  ...,
  mode = c("light", "dark"),
  base = ggplot2::theme_bw()
)
}
\arguments{
\item{theme}{Name of the theme to use. One of "ghibli_mononoke", "islamic_samarquand",
"pomological_green", "pomological_red", "nord_blue", "swiss_red", "doc" or "oc_exams"}

\item{main}{One of `main`, `main_latex` or `main_exa`}

\item{plot}{One of `plot`, `plot_latex` or `plot_exa`}

\item{mode}{One of `light` or `dark`}

\item{base}{A ggplot2 theme}
}
\value{
An object of class \code{\link[ggplot2]{theme}()}.
}
\description{
ggeo theme
}
