% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharter_themes.R
\name{hc_purple_theme}
\alias{hc_purple_theme}
\alias{hc_samarqand_theme}
\alias{hc_jerusalem_theme}
\alias{hc_web_theme}
\title{Highcharts themes}
\usage{
hc_purple_theme(hc)

hc_samarqand_theme(hc)

hc_jerusalem_theme(hc)

hc_web_theme(hc)
}
\arguments{
\item{hc}{A highcharts object}
}
\value{
A highcharts object
}
\description{
These functions can be used to modify the theme of
a highcharts plot.
}
\examples{
highcharter::hchart(
  cars, "point",
  highcharter::hcaes(x = speed, y = dist)
) |>
  hc_purple_theme()

highcharter::hchart(
  cars, "point",
  highcharter::hcaes(x = speed, y = dist)
) |>
  hc_samarqand_theme()

highcharter::hchart(
  cars, "point",
  highcharter::hcaes(x = speed, y = dist)
) |>
  hc_web_theme()

}
